backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
selenium==4.27.1
python-dotenv==1.0.1
pymongo[snappy,gssapi,srv,tls]
dnspython==2.7.0
requests
django-cors-headers
